testing backchannel lookup of sid

+++ lambdas/handlers/back_channel_logout_handler.py
@@ -1,6 +1,7 @@
 import os
 from urllib.parse import parse_qs
 
+from boto3.dynamodb.conditions import Attr
 from botocore.exceptions import ClientError
 from enums.logging_app_interaction import LoggingAppInteraction
 from services.dynamo_service import DynamoDBService
@@ -40,8 +41,8 @@ def logout_handler(token):
     try:
         oidc_service = OidcService()
         decoded_token = oidc_service.validate_and_decode_token(token)
-        session_id = decoded_token["sid"]
-        remove_session_from_dynamo_db(session_id)
+        sid = decoded_token["sid"]
+        remove_session_from_dynamo_db(sid)
 
     except ClientError as e:
         logger.error(f"Error logging out user: {e}", {"Result": "Unsuccessful logout"})
@@ -66,10 +67,23 @@ def logout_handler(token):
     return ApiGatewayResponse(200, "", "POST").create_api_gateway_response()
 
 
-def remove_session_from_dynamo_db(session_id):
+def remove_session_from_dynamo_db(sid):
     dynamodb_name = os.environ["AUTH_DYNAMODB_NAME"]
     dynamodb_service = DynamoDBService()
-    dynamodb_service.delete_item(
-        key={"NDRSessionId": session_id}, table_name=dynamodb_name
+    
+    filter_sid = Attr("sid").eq(sid)
+    db_response = dynamodb_service.scan_table(
+        dynamodb_name, filter_expression=filter_sid
     )
-    logger.info(f"Session removed: {session_id}", {"Result": "Successful logout"})
+
+    logger.info(f'db response {db_response}')
+    if "Items" in db_response:
+        items = db_response["Items"]
+        
+        while "NDRSessionId" in items:
+            logger.info(f'ndr session id {db_response["NDRSessionId"]}')
+            dynamodb_service.delete_item(
+                key={"NDRSessionId": db_response["NDRSessionId"]}, table_name=dynamodb_name
+            )
+
+    logger.info(f"Session removed: {sid}", {"Result": "Successful logout"})
